+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor changes
 * Get ACL link on container creation to avoid unnecessary HEAD request.
+* Fix issue attempting to get ACL link when resource creation failed.
 
 ## Release 1.1.1